Refactored fiftyone ingest
Using pandas dataframes as a common interface for uploading fo datasets
seems prudent. Experiment loaders for which the constituent data exists
as image files on disk can be rewritten to match the CQ1 exeperiment
loader. Integrating with fo is then straightforward.

Experiment loaders that rely on chunked array formats are trickier.
While exporting a PNG view of the experiment is easy, accessing the
underlying raw data for quantitative purposes is less so.  My current
thinking is to export such experiments as a series of TIFFs to provide
a consistent file-based interface for analyses that use fiftyone.