Tweak survival + CQ1 loading
We started getting random lost acquisitions while loading CQ1 data. This
created a need for handling missing data in survival analysis. We
modified the CQ1 loader to output NaNs for missing data, and added
handling to the survival analysis to ignore those frames.

We also modified the survival analysis to analyze the experiment on the
field-series level, rather than on a timeseries level. This allows the
analysis to scale better (assuming that the number of fields per well
does not scale too much,) handling long timeseries data without blowing
up. In the future, we may need to select a more constrained block size
if the number of fields per well varies exeperiment-to-experiment as
much as the length of timeseries.

It's important to note that survival should be run with the
threads_per_worker config option set to 1, as there appears to be a
memory leak in stardist (at least on my machine, which admittedly has
other quirks with newer versions of stardist that aren't present on
other machines I've tested.)