Add simplified upload script for CVAT
This script is a refactor of the original. At least to me, it's much
easier to make sense of. The main functional difference is metadata is
no longer stored in task names- we make a record of the uploaded
annotations in an experiment's results directory, then use that record
(a csv file) to determine the correspondence between the underlying raw
data and any annotations coming out of CVAT.