Rename IO wrappers
We avoid importing business logic into click module namespaces to keep
load times short. As a result, we have to push IO and functional logic
to a different module. I have been separating IO and functional logic
somewhat, but this commit makes that separation more explicit. The
functional component of the business logic is named plainly. For instance,
the functional component of nuc-cyto business logic is named `nuc_cyto.`
The corresponding wrappers that handle IO and invoke the functional
business logic are named `do_X` where `X` is the name of the wrapped
function.

Previously, I used `handle_X` to denote the functional business logic.
I don't really know why; the context in which I see `handle_X` used is
typically in the context of event handling.