change 'field' values to integers
'field' values for legacy experiments were tuples (0_0, 0_1, ...), which
are inconsistent with other high-throughput experiment loaders (CQ1.)

The benefit of using tuples for field values is the imaging order is
baked-in, so a stitching algorithm can operate based on the coordinate
data without having to calculate the tile positions given an imaging
order. This is a nice feature, but right now we're not doing much
stitching.

Eventually, it would be nice to have a standardized format for xarray
experiments, along with smarter zarr caching. For now, we'll just revert
field values to integers to be consistent.